+++ model_navigator/utils/cli.py
@@ -407,7 +407,7 @@ def common_options(f):
                 "NVIDIA framework and Triton container version to use "
                 "(refer to https://docs.nvidia.com/deeplearning/frameworks/support-matrix/index.html and https://docs.nvidia.com/deeplearning/triton-inference-server/release-notes/index.html for details)."
             ),
-            default="21.05",
+            default="21.06",
         ),
         click.option(
             "--gpus",